+++ tests/_cases/template_engines/index.php
@@ -1,0 +1,5 @@
+<?php
+require_once "platesphp/Report.php";
+
+$report = new Report;
+$report->run()->render("report");